requirements.txt
python-dotenv
langchain==0.2.1
langchain-core==0.2.2
langchain-community==0.2.1
langchain-google-genai==1.0.5
pinecone-client==3.2.2
langchain-pinecone==0.1.1
langchain-openai==0.1.8
chainlit==1.1.202
langsmith==0.1.63
openai==1.30.5
PyMuPDF
python-dotenv
flask
langchain
pinecone-client